✨ feat(extensions): convert todos to directory layout with tests
