+++ extensions/todos/components.ts
@@ -1,0 +1,572 @@
+/**
+ * TUI components for the todos extension.
+ */
+import path from "node:path";
+import { DynamicBorder, copyToClipboard, getMarkdownTheme, type Theme } from "@mariozechner/pi-coding-agent";
+import {
+	Container,
+	type Focusable,
+	Input,
+	Key,
+	Markdown,
+	SelectList,
+	Spacer,
+	type SelectItem,
+	Text,
+	TUI,
+	matchesKey,
+	truncateToWidth,
+	visibleWidth,
+	type Keybinding,
+} from "@mariozechner/pi-tui";
+import type { TodoFrontMatter, TodoMenuAction, TodoOverlayAction, TodoRecord } from "./types.ts";
+import {
+	filterTodos,
+	formatTodoId,
+	isTodoClosed,
+	renderAssignmentSuffix,
+	renderTodoHeading,
+} from "./helpers.ts";
+import { getTodoPath } from "./helpers.ts";
+
+type KeybindingMatcher = {
+	matches: (keyData: string, keybindingId: Keybinding) => boolean;
+};
+
+// ─── TodoSelectorComponent ────────────────────────────────────────────────────
+
+export class TodoSelectorComponent extends Container implements Focusable {
+	private searchInput: Input;
+	private listContainer: Container;
+	private allTodos: TodoFrontMatter[];
+	private filteredTodos: TodoFrontMatter[];
+	private selectedIndex = 0;
+	private onSelectCallback: (todo: TodoFrontMatter) => void;
+	private onCancelCallback: () => void;
+	private tui: TUI;
+	private theme: Theme;
+	private keybindings: KeybindingMatcher;
+	private headerText: Text;
+	private hintText: Text;
+	private currentSessionId?: string;
+
+	private _focused = false;
+	get focused(): boolean {
+		return this._focused;
+	}
+	set focused(value: boolean) {
+		this._focused = value;
+		this.searchInput.focused = value;
+	}
+
+	constructor(
+		tui: TUI,
+		theme: Theme,
+		keybindings: KeybindingMatcher,
+		todos: TodoFrontMatter[],
+		onSelect: (todo: TodoFrontMatter) => void,
+		onCancel: () => void,
+		initialSearchInput?: string,
+		currentSessionId?: string,
+		private onQuickAction?: (todo: TodoFrontMatter, action: "work" | "refine") => void,
+	) {
+		super();
+		this.tui = tui;
+		this.theme = theme;
+		this.keybindings = keybindings;
+		this.currentSessionId = currentSessionId;
+		this.allTodos = todos;
+		this.filteredTodos = todos;
+		this.onSelectCallback = onSelect;
+		this.onCancelCallback = onCancel;
+
+		this.addChild(new DynamicBorder((s: string) => theme.fg("accent", s)));
+		this.addChild(new Spacer(1));
+
+		this.headerText = new Text("", 1, 0);
+		this.addChild(this.headerText);
+		this.addChild(new Spacer(1));
+
+		this.searchInput = new Input();
+		if (initialSearchInput) {
+			this.searchInput.setValue(initialSearchInput);
+		}
+		this.searchInput.onSubmit = () => {
+			const selected = this.filteredTodos[this.selectedIndex];
+			if (selected) this.onSelectCallback(selected);
+		};
+		this.addChild(this.searchInput);
+
+		this.addChild(new Spacer(1));
+		this.listContainer = new Container();
+		this.addChild(this.listContainer);
+
+		this.addChild(new Spacer(1));
+		this.hintText = new Text("", 1, 0);
+		this.addChild(this.hintText);
+		this.addChild(new Spacer(1));
+		this.addChild(new DynamicBorder((s: string) => theme.fg("accent", s)));
+
+		this.updateHeader();
+		this.updateHints();
+		this.applyFilter(this.searchInput.getValue());
+	}
+
+	setTodos(todos: TodoFrontMatter[]): void {
+		this.allTodos = todos;
+		this.updateHeader();
+		this.applyFilter(this.searchInput.getValue());
+		this.tui.requestRender();
+	}
+
+	getSearchValue(): string {
+		return this.searchInput.getValue();
+	}
+
+	private updateHeader(): void {
+		const openCount = this.allTodos.filter((todo) => !isTodoClosed(todo.status)).length;
+		const closedCount = this.allTodos.length - openCount;
+		const title = `Todos (${openCount} open, ${closedCount} closed)`;
+		this.headerText.setText(this.theme.fg("accent", this.theme.bold(title)));
+	}
+
+	private updateHints(): void {
+		this.hintText.setText(
+			this.theme.fg(
+				"dim",
+				"Type to search • ↑↓ select • Enter actions • Ctrl+Shift+W work • Ctrl+Shift+R refine • Esc close",
+			),
+		);
+	}
+
+	private applyFilter(query: string): void {
+		this.filteredTodos = filterTodos(this.allTodos, query);
+		this.selectedIndex = Math.min(
+			this.selectedIndex,
+			Math.max(0, this.filteredTodos.length - 1),
+		);
+		this.updateList();
+	}
+
+	private updateList(): void {
+		this.listContainer.clear();
+
+		if (this.filteredTodos.length === 0) {
+			this.listContainer.addChild(
+				new Text(this.theme.fg("muted", "  No matching todos"), 0, 0),
+			);
+			return;
+		}
+
+		const maxVisible = 10;
+		const startIndex = Math.max(
+			0,
+			Math.min(
+				this.selectedIndex - Math.floor(maxVisible / 2),
+				this.filteredTodos.length - maxVisible,
+			),
+		);
+		const endIndex = Math.min(startIndex + maxVisible, this.filteredTodos.length);
+
+		for (let i = startIndex; i < endIndex; i += 1) {
+			const todo = this.filteredTodos[i];
+			if (!todo) continue;
+			const isSelected = i === this.selectedIndex;
+			const closed = isTodoClosed(todo.status);
+			const prefix = isSelected ? this.theme.fg("accent", "→ ") : "  ";
+			const titleColor = isSelected ? "accent" : closed ? "dim" : "text";
+			const statusColor = closed ? "dim" : "success";
+			const tagText = todo.tags.length ? ` [${todo.tags.join(", ")}]` : "";
+			const assignmentText = renderAssignmentSuffix(this.theme, todo, this.currentSessionId);
+			const line =
+				prefix +
+				this.theme.fg("accent", formatTodoId(todo.id)) +
+				" " +
+				this.theme.fg(titleColor, todo.title || "(untitled)") +
+				this.theme.fg("muted", tagText) +
+				assignmentText +
+				" " +
+				this.theme.fg(statusColor, `(${todo.status || "open"})`);
+			this.listContainer.addChild(new Text(line, 0, 0));
+		}
+
+		if (startIndex > 0 || endIndex < this.filteredTodos.length) {
+			const scrollInfo = this.theme.fg(
+				"dim",
+				`  (${this.selectedIndex + 1}/${this.filteredTodos.length})`,
+			);
+			this.listContainer.addChild(new Text(scrollInfo, 0, 0));
+		}
+	}
+
+	handleInput(keyData: string): void {
+		const kb = this.keybindings;
+		if (kb.matches(keyData, "tui.select.up")) {
+			if (this.filteredTodos.length === 0) return;
+			this.selectedIndex =
+				this.selectedIndex === 0
+					? this.filteredTodos.length - 1
+					: this.selectedIndex - 1;
+			this.updateList();
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.down")) {
+			if (this.filteredTodos.length === 0) return;
+			this.selectedIndex =
+				this.selectedIndex === this.filteredTodos.length - 1
+					? 0
+					: this.selectedIndex + 1;
+			this.updateList();
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.confirm")) {
+			const selected = this.filteredTodos[this.selectedIndex];
+			if (selected) this.onSelectCallback(selected);
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.cancel")) {
+			this.onCancelCallback();
+			return;
+		}
+		if (matchesKey(keyData, Key.ctrlShift("r"))) {
+			const selected = this.filteredTodos[this.selectedIndex];
+			if (selected && this.onQuickAction) this.onQuickAction(selected, "refine");
+			return;
+		}
+		if (matchesKey(keyData, Key.ctrlShift("w"))) {
+			const selected = this.filteredTodos[this.selectedIndex];
+			if (selected && this.onQuickAction) this.onQuickAction(selected, "work");
+			return;
+		}
+		this.searchInput.handleInput(keyData);
+		this.applyFilter(this.searchInput.getValue());
+	}
+
+	override invalidate(): void {
+		super.invalidate();
+		this.updateHeader();
+		this.updateHints();
+		this.updateList();
+	}
+}
+
+// ─── TodoActionMenuComponent ──────────────────────────────────────────────────
+
+export class TodoActionMenuComponent extends Container {
+	private selectList: SelectList;
+	private onSelectCallback: (action: TodoMenuAction) => void;
+	private onCancelCallback: () => void;
+
+	constructor(
+		theme: Theme,
+		todo: TodoRecord,
+		onSelect: (action: TodoMenuAction) => void,
+		onCancel: () => void,
+	) {
+		super();
+		this.onSelectCallback = onSelect;
+		this.onCancelCallback = onCancel;
+
+		const closed = isTodoClosed(todo.status);
+		const title = todo.title || "(untitled)";
+		const options: SelectItem[] = [
+			{ value: "view", label: "view", description: "View todo" },
+			{ value: "work", label: "work", description: "Work on todo" },
+			{ value: "refine", label: "refine", description: "Refine task" },
+			...(closed
+				? [{ value: "reopen", label: "reopen", description: "Reopen todo" }]
+				: [{ value: "close", label: "close", description: "Close todo" }]),
+			...(todo.assigned_to_session
+				? [{ value: "release", label: "release", description: "Release assignment" }]
+				: []),
+			{
+				value: "copyPath",
+				label: "copy path",
+				description: "Copy absolute path to clipboard",
+			},
+			{
+				value: "copyText",
+				label: "copy text",
+				description: "Copy title and body to clipboard",
+			},
+			{ value: "delete", label: "delete", description: "Delete todo" },
+		];
+
+		this.addChild(new DynamicBorder((s: string) => theme.fg("accent", s)));
+		this.addChild(
+			new Text(
+				theme.fg("accent", theme.bold(`Actions for ${formatTodoId(todo.id)} "${title}"`)),
+			),
+		);
+
+		this.selectList = new SelectList(options, options.length, {
+			selectedPrefix: (text) => theme.fg("accent", text),
+			selectedText: (text) => theme.fg("accent", text),
+			description: (text) => theme.fg("muted", text),
+			scrollInfo: (text) => theme.fg("dim", text),
+			noMatch: (text) => theme.fg("warning", text),
+		});
+
+		this.selectList.onSelect = (item) =>
+			this.onSelectCallback(item.value as TodoMenuAction);
+		this.selectList.onCancel = () => this.onCancelCallback();
+
+		this.addChild(this.selectList);
+		this.addChild(new Text(theme.fg("dim", "Enter to confirm • Esc back")));
+		this.addChild(new DynamicBorder((s: string) => theme.fg("accent", s)));
+	}
+
+	handleInput(keyData: string): void {
+		this.selectList.handleInput(keyData);
+	}
+
+	override invalidate(): void {
+		super.invalidate();
+	}
+}
+
+// ─── TodoDeleteConfirmComponent ───────────────────────────────────────────────
+
+export class TodoDeleteConfirmComponent extends Container {
+	private selectList: SelectList;
+	private onConfirm: (confirmed: boolean) => void;
+
+	constructor(theme: Theme, message: string, onConfirm: (confirmed: boolean) => void) {
+		super();
+		this.onConfirm = onConfirm;
+
+		const options: SelectItem[] = [
+			{ value: "yes", label: "Yes" },
+			{ value: "no", label: "No" },
+		];
+
+		this.addChild(new DynamicBorder((s: string) => theme.fg("accent", s)));
+		this.addChild(new Text(theme.fg("accent", message)));
+
+		this.selectList = new SelectList(options, options.length, {
+			selectedPrefix: (text) => theme.fg("accent", text),
+			selectedText: (text) => theme.fg("accent", text),
+			description: (text) => theme.fg("muted", text),
+			scrollInfo: (text) => theme.fg("dim", text),
+			noMatch: (text) => theme.fg("warning", text),
+		});
+
+		this.selectList.onSelect = (item) => this.onConfirm(item.value === "yes");
+		this.selectList.onCancel = () => this.onConfirm(false);
+
+		this.addChild(this.selectList);
+		this.addChild(new Text(theme.fg("dim", "Enter to confirm • Esc back")));
+		this.addChild(new DynamicBorder((s: string) => theme.fg("accent", s)));
+	}
+
+	handleInput(keyData: string): void {
+		this.selectList.handleInput(keyData);
+	}
+
+	override invalidate(): void {
+		super.invalidate();
+	}
+}
+
+// ─── TodoDetailOverlayComponent ───────────────────────────────────────────────
+
+export class TodoDetailOverlayComponent {
+	private todo: TodoRecord;
+	private theme: Theme;
+	private tui: TUI;
+	private markdown: Markdown;
+	private scrollOffset = 0;
+	private viewHeight = 0;
+	private totalLines = 0;
+	private onAction: (action: TodoOverlayAction) => void;
+	private keybindings: KeybindingMatcher;
+
+	constructor(
+		tui: TUI,
+		theme: Theme,
+		keybindings: KeybindingMatcher,
+		todo: TodoRecord,
+		onAction: (action: TodoOverlayAction) => void,
+	) {
+		this.tui = tui;
+		this.theme = theme;
+		this.keybindings = keybindings;
+		this.todo = todo;
+		this.onAction = onAction;
+		this.markdown = new Markdown(this.getMarkdownText(), 1, 0, getMarkdownTheme());
+	}
+
+	private getMarkdownText(): string {
+		const body = this.todo.body?.trim();
+		return body ? body : "_No details yet._";
+	}
+
+	handleInput(keyData: string): void {
+		const kb = this.keybindings;
+		if (kb.matches(keyData, "tui.select.cancel")) {
+			this.onAction("back");
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.confirm")) {
+			this.onAction("work");
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.up")) {
+			this.scrollBy(-1);
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.down")) {
+			this.scrollBy(1);
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.pageUp") || matchesKey(keyData, Key.left)) {
+			this.scrollBy(-this.viewHeight || -1);
+			return;
+		}
+		if (kb.matches(keyData, "tui.select.pageDown") || matchesKey(keyData, Key.right)) {
+			this.scrollBy(this.viewHeight || 1);
+			return;
+		}
+	}
+
+	render(width: number): string[] {
+		const maxHeight = this.getMaxHeight();
+		const headerLines = 3;
+		const footerLines = 3;
+		const borderLines = 2;
+		const innerWidth = Math.max(10, width - 2);
+		const contentHeight = Math.max(1, maxHeight - headerLines - footerLines - borderLines);
+
+		const markdownLines = this.markdown.render(innerWidth);
+		this.totalLines = markdownLines.length;
+		this.viewHeight = contentHeight;
+		const maxScroll = Math.max(0, this.totalLines - contentHeight);
+		this.scrollOffset = Math.max(0, Math.min(this.scrollOffset, maxScroll));
+
+		const visibleLines = markdownLines.slice(
+			this.scrollOffset,
+			this.scrollOffset + contentHeight,
+		);
+		const lines: string[] = [];
+
+		lines.push(this.buildTitleLine(innerWidth));
+		lines.push(this.buildMetaLine(innerWidth));
+		lines.push("");
+
+		for (const line of visibleLines) {
+			lines.push(truncateToWidth(line, innerWidth));
+		}
+		while (lines.length < headerLines + contentHeight) {
+			lines.push("");
+		}
+
+		lines.push("");
+		lines.push(this.buildActionLine(innerWidth));
+
+		const borderColor = (text: string) => this.theme.fg("borderMuted", text);
+		const top = borderColor(`┌${"─".repeat(innerWidth)}┐`);
+		const bottom = borderColor(`└${"─".repeat(innerWidth)}┘`);
+		const framedLines = lines.map((line) => {
+			const truncated = truncateToWidth(line, innerWidth);
+			const padding = Math.max(0, innerWidth - visibleWidth(truncated));
+			return borderColor("│") + truncated + " ".repeat(padding) + borderColor("│");
+		});
+
+		return [top, ...framedLines, bottom].map((line) => truncateToWidth(line, width));
+	}
+
+	invalidate(): void {
+		this.markdown = new Markdown(this.getMarkdownText(), 1, 0, getMarkdownTheme());
+	}
+
+	private getMaxHeight(): number {
+		const rows = this.tui.terminal.rows || 24;
+		return Math.max(10, Math.floor(rows * 0.8));
+	}
+
+	private buildTitleLine(width: number): string {
+		const titleText = this.todo.title
+			? ` ${this.todo.title} `
+			: ` Todo ${formatTodoId(this.todo.id)} `;
+		const titleWidth = visibleWidth(titleText);
+		if (titleWidth >= width) {
+			return truncateToWidth(this.theme.fg("accent", titleText.trim()), width);
+		}
+		const leftWidth = Math.max(0, Math.floor((width - titleWidth) / 2));
+		const rightWidth = Math.max(0, width - titleWidth - leftWidth);
+		return (
+			this.theme.fg("borderMuted", "─".repeat(leftWidth)) +
+			this.theme.fg("accent", titleText) +
+			this.theme.fg("borderMuted", "─".repeat(rightWidth))
+		);
+	}
+
+	private buildMetaLine(width: number): string {
+		const status = this.todo.status || "open";
+		const statusColor = isTodoClosed(status) ? "dim" : "success";
+		const tagText = this.todo.tags.length ? this.todo.tags.join(", ") : "no tags";
+		const line =
+			this.theme.fg("accent", formatTodoId(this.todo.id)) +
+			this.theme.fg("muted", " • ") +
+			this.theme.fg(statusColor, status) +
+			this.theme.fg("muted", " • ") +
+			this.theme.fg("muted", tagText);
+		return truncateToWidth(line, width);
+	}
+
+	private buildActionLine(width: number): string {
+		const work = this.theme.fg("accent", "enter") + this.theme.fg("muted", " work on todo");
+		const back = this.theme.fg("dim", "esc back");
+		const nav = this.theme.fg("dim", "↑/↓: move. ←/→: page.");
+		const pieces = [work, back, nav];
+
+		let line = pieces.join(this.theme.fg("muted", " • "));
+		if (this.totalLines > this.viewHeight) {
+			const start = Math.min(this.totalLines, this.scrollOffset + 1);
+			const end = Math.min(this.totalLines, this.scrollOffset + this.viewHeight);
+			const scrollInfo = this.theme.fg("dim", ` ${start}-${end}/${this.totalLines}`);
+			line += scrollInfo;
+		}
+
+		return truncateToWidth(line, width);
+	}
+
+	private scrollBy(delta: number): void {
+		const maxScroll = Math.max(0, this.totalLines - this.viewHeight);
+		this.scrollOffset = Math.max(0, Math.min(this.scrollOffset + delta, maxScroll));
+	}
+}
+
+// ─── Clipboard helpers ────────────────────────────────────────────────────────
+
+export function copyTodoPathToClipboard(
+	todosDir: string,
+	todoId: string,
+	notify: (message: string, type: "info" | "error") => void,
+): void {
+	const filePath = getTodoPath(todosDir, todoId);
+	const absolutePath = path.resolve(filePath);
+	try {
+		copyToClipboard(absolutePath);
+		notify(`Copied ${absolutePath} to clipboard`, "info");
+	} catch (error) {
+		const message = error instanceof Error ? error.message : String(error);
+		notify(message, "error");
+	}
+}
+
+export function copyTodoTextToClipboard(
+	record: TodoRecord,
+	notify: (message: string, type: "info" | "error") => void,
+): void {
+	const title = record.title || "(untitled)";
+	const body = record.body?.trim() || "";
+	const text = body ? `# ${title}\n\n${body}` : `# ${title}`;
+	try {
+		copyToClipboard(text);
+		notify("Copied todo text to clipboard", "info");
+	} catch (error) {
+		const message = error instanceof Error ? error.message : String(error);
+		notify(message, "error");
+	}
+}
