+++ extensions/todos/helpers.ts
@@ -1,0 +1,334 @@
+/**
+ * Pure helper functions for todos — formatting, serialization, sorting.
+ * No I/O or side effects; safe to import in tests.
+ */
+import path from "node:path";
+import type { Theme } from "@mariozechner/pi-coding-agent";
+import { keyHint } from "@mariozechner/pi-coding-agent";
+import { fuzzyMatch } from "@mariozechner/pi-tui";
+import {
+	TODO_ID_PREFIX,
+	TODO_ID_PATTERN,
+	TODO_DIR_NAME,
+	TODO_PATH_ENV,
+	type TodoFrontMatter,
+	type TodoRecord,
+} from "./types.ts";
+
+// ─── ID helpers ───────────────────────────────────────────────────────────────
+
+export function formatTodoId(id: string): string {
+	return `${TODO_ID_PREFIX}${id}`;
+}
+
+export function normalizeTodoId(id: string): string {
+	let trimmed = id.trim();
+	if (trimmed.startsWith("#")) {
+		trimmed = trimmed.slice(1);
+	}
+	if (trimmed.toUpperCase().startsWith(TODO_ID_PREFIX)) {
+		trimmed = trimmed.slice(TODO_ID_PREFIX.length);
+	}
+	return trimmed;
+}
+
+export function validateTodoId(id: string): { id: string } | { error: string } {
+	const normalized = normalizeTodoId(id);
+	if (!normalized || !TODO_ID_PATTERN.test(normalized)) {
+		return { error: "Invalid todo id. Expected TODO-<hex>." };
+	}
+	return { id: normalized.toLowerCase() };
+}
+
+export function displayTodoId(id: string): string {
+	return formatTodoId(normalizeTodoId(id));
+}
+
+// ─── Status helpers ───────────────────────────────────────────────────────────
+
+export function isTodoClosed(status: string): boolean {
+	return ["closed", "done"].includes(status.toLowerCase());
+}
+
+export function clearAssignmentIfClosed(todo: TodoFrontMatter): void {
+	if (isTodoClosed(getTodoStatus(todo))) {
+		todo.assigned_to_session = undefined;
+	}
+}
+
+export function getTodoTitle(todo: TodoFrontMatter): string {
+	return todo.title || "(untitled)";
+}
+
+export function getTodoStatus(todo: TodoFrontMatter): string {
+	return todo.status || "open";
+}
+
+// ─── Sorting & filtering ──────────────────────────────────────────────────────
+
+export function sortTodos(todos: TodoFrontMatter[]): TodoFrontMatter[] {
+	return [...todos].sort((a, b) => {
+		const aClosed = isTodoClosed(a.status);
+		const bClosed = isTodoClosed(b.status);
+		if (aClosed !== bClosed) return aClosed ? 1 : -1;
+		const aAssigned = !aClosed && Boolean(a.assigned_to_session);
+		const bAssigned = !bClosed && Boolean(b.assigned_to_session);
+		if (aAssigned !== bAssigned) return aAssigned ? -1 : 1;
+		return (a.created_at || "").localeCompare(b.created_at || "");
+	});
+}
+
+function buildTodoSearchText(todo: TodoFrontMatter): string {
+	const tags = todo.tags.join(" ");
+	const assignment = todo.assigned_to_session ? `assigned:${todo.assigned_to_session}` : "";
+	return `${formatTodoId(todo.id)} ${todo.id} ${todo.title} ${tags} ${todo.status} ${assignment}`.trim();
+}
+
+export function filterTodos(todos: TodoFrontMatter[], query: string): TodoFrontMatter[] {
+	const trimmed = query.trim();
+	if (!trimmed) return todos;
+
+	const tokens = trimmed
+		.split(/\s+/)
+		.map((token) => token.trim())
+		.filter(Boolean);
+
+	if (tokens.length === 0) return todos;
+
+	const matches: Array<{ todo: TodoFrontMatter; score: number }> = [];
+	for (const todo of todos) {
+		const text = buildTodoSearchText(todo);
+		let totalScore = 0;
+		let matched = true;
+		for (const token of tokens) {
+			const result = fuzzyMatch(token, text);
+			if (!result.matches) {
+				matched = false;
+				break;
+			}
+			totalScore += result.score;
+		}
+		if (matched) {
+			matches.push({ todo, score: totalScore });
+		}
+	}
+
+	return matches
+		.sort((a, b) => {
+			const aClosed = isTodoClosed(a.todo.status);
+			const bClosed = isTodoClosed(b.todo.status);
+			if (aClosed !== bClosed) return aClosed ? 1 : -1;
+			const aAssigned = !aClosed && Boolean(a.todo.assigned_to_session);
+			const bAssigned = !bClosed && Boolean(b.todo.assigned_to_session);
+			if (aAssigned !== bAssigned) return aAssigned ? -1 : 1;
+			return a.score - b.score;
+		})
+		.map((match) => match.todo);
+}
+
+// ─── Path helpers ─────────────────────────────────────────────────────────────
+
+export function getTodosDir(cwd: string): string {
+	const overridePath = process.env[TODO_PATH_ENV];
+	if (overridePath && overridePath.trim()) {
+		return path.resolve(cwd, overridePath.trim());
+	}
+	return path.resolve(cwd, TODO_DIR_NAME);
+}
+
+export function getTodosDirLabel(cwd: string): string {
+	const overridePath = process.env[TODO_PATH_ENV];
+	if (overridePath && overridePath.trim()) {
+		return path.resolve(cwd, overridePath.trim());
+	}
+	return TODO_DIR_NAME;
+}
+
+export function getTodoPath(todosDir: string, id: string): string {
+	return path.join(todosDir, `${id}.md`);
+}
+
+export function getLockPath(todosDir: string, id: string): string {
+	return path.join(todosDir, `${id}.lock`);
+}
+
+// ─── Formatting helpers ───────────────────────────────────────────────────────
+
+export function formatAssignmentSuffix(todo: TodoFrontMatter): string {
+	return todo.assigned_to_session ? ` (assigned: ${todo.assigned_to_session})` : "";
+}
+
+export function formatTodoHeading(todo: TodoFrontMatter): string {
+	const tagText = todo.tags.length ? ` [${todo.tags.join(", ")}]` : "";
+	return `${formatTodoId(todo.id)} ${getTodoTitle(todo)}${tagText}${formatAssignmentSuffix(todo)}`;
+}
+
+export function splitTodosByAssignment(todos: TodoFrontMatter[]): {
+	assignedTodos: TodoFrontMatter[];
+	openTodos: TodoFrontMatter[];
+	closedTodos: TodoFrontMatter[];
+} {
+	const assignedTodos: TodoFrontMatter[] = [];
+	const openTodos: TodoFrontMatter[] = [];
+	const closedTodos: TodoFrontMatter[] = [];
+	for (const todo of todos) {
+		if (isTodoClosed(getTodoStatus(todo))) {
+			closedTodos.push(todo);
+			continue;
+		}
+		if (todo.assigned_to_session) {
+			assignedTodos.push(todo);
+		} else {
+			openTodos.push(todo);
+		}
+	}
+	return { assignedTodos, openTodos, closedTodos };
+}
+
+export function formatTodoList(todos: TodoFrontMatter[]): string {
+	if (!todos.length) return "No todos.";
+
+	const { assignedTodos, openTodos, closedTodos } = splitTodosByAssignment(todos);
+	const lines: string[] = [];
+	const pushSection = (label: string, sectionTodos: TodoFrontMatter[]) => {
+		lines.push(`${label} (${sectionTodos.length}):`);
+		if (!sectionTodos.length) {
+			lines.push("  none");
+			return;
+		}
+		for (const todo of sectionTodos) {
+			lines.push(`  ${formatTodoHeading(todo)}`);
+		}
+	};
+
+	pushSection("Assigned todos", assignedTodos);
+	pushSection("Open todos", openTodos);
+	pushSection("Closed todos", closedTodos);
+	return lines.join("\n");
+}
+
+export function buildRefinePrompt(todoId: string, title: string): string {
+	return (
+		`let's refine task ${formatTodoId(todoId)} "${title}": ` +
+		"Ask me for the missing details needed to refine the todo together. Do not rewrite the todo yet and do not make assumptions. " +
+		"Ask clear, concrete questions and wait for my answers before drafting any structured description.\n\n"
+	);
+}
+
+// ─── Serialization for agent ──────────────────────────────────────────────────
+
+export function serializeTodoForAgent(todo: TodoRecord): string {
+	const payload = { ...todo, id: formatTodoId(todo.id) };
+	return JSON.stringify(payload, null, 2);
+}
+
+export function serializeTodoListForAgent(todos: TodoFrontMatter[]): string {
+	const { assignedTodos, openTodos, closedTodos } = splitTodosByAssignment(todos);
+	const mapTodo = (todo: TodoFrontMatter) => ({ ...todo, id: formatTodoId(todo.id) });
+	return JSON.stringify(
+		{
+			assigned: assignedTodos.map(mapTodo),
+			open: openTodos.map(mapTodo),
+			closed: closedTodos.map(mapTodo),
+		},
+		null,
+		2,
+	);
+}
+
+// ─── Render helpers (require theme) ──────────────────────────────────────────
+
+export function renderAssignmentSuffix(
+	theme: Theme,
+	todo: TodoFrontMatter,
+	currentSessionId?: string,
+): string {
+	if (!todo.assigned_to_session) return "";
+	const isCurrent = todo.assigned_to_session === currentSessionId;
+	const color = isCurrent ? "success" : "dim";
+	const suffix = isCurrent ? ", current" : "";
+	return theme.fg(color, ` (assigned: ${todo.assigned_to_session}${suffix})`);
+}
+
+export function renderTodoHeading(
+	theme: Theme,
+	todo: TodoFrontMatter,
+	currentSessionId?: string,
+): string {
+	const closed = isTodoClosed(getTodoStatus(todo));
+	const titleColor = closed ? "dim" : "text";
+	const tagText = todo.tags.length ? theme.fg("dim", ` [${todo.tags.join(", ")}]`) : "";
+	const assignmentText = renderAssignmentSuffix(theme, todo, currentSessionId);
+	return (
+		theme.fg("accent", formatTodoId(todo.id)) +
+		" " +
+		theme.fg(titleColor, getTodoTitle(todo)) +
+		tagText +
+		assignmentText
+	);
+}
+
+export function renderTodoList(
+	theme: Theme,
+	todos: TodoFrontMatter[],
+	expanded: boolean,
+	currentSessionId?: string,
+): string {
+	if (!todos.length) return theme.fg("dim", "No todos");
+
+	const { assignedTodos, openTodos, closedTodos } = splitTodosByAssignment(todos);
+	const lines: string[] = [];
+	const pushSection = (label: string, sectionTodos: TodoFrontMatter[]) => {
+		lines.push(theme.fg("muted", `${label} (${sectionTodos.length})`));
+		if (!sectionTodos.length) {
+			lines.push(theme.fg("dim", "  none"));
+			return;
+		}
+		const maxItems = expanded ? sectionTodos.length : Math.min(sectionTodos.length, 3);
+		for (let i = 0; i < maxItems; i++) {
+			lines.push(`  ${renderTodoHeading(theme, sectionTodos[i], currentSessionId)}`);
+		}
+		if (!expanded && sectionTodos.length > maxItems) {
+			lines.push(theme.fg("dim", `  ... ${sectionTodos.length - maxItems} more`));
+		}
+	};
+
+	const sections: Array<{ label: string; todos: TodoFrontMatter[] }> = [
+		{ label: "Assigned todos", todos: assignedTodos },
+		{ label: "Open todos", todos: openTodos },
+		{ label: "Closed todos", todos: closedTodos },
+	];
+
+	sections.forEach((section, index) => {
+		if (index > 0) lines.push("");
+		pushSection(section.label, section.todos);
+	});
+
+	return lines.join("\n");
+}
+
+export function renderTodoDetail(theme: Theme, todo: TodoRecord, expanded: boolean): string {
+	const summary = renderTodoHeading(theme, todo);
+	if (!expanded) return summary;
+
+	const tags = todo.tags.length ? todo.tags.join(", ") : "none";
+	const createdAt = todo.created_at || "unknown";
+	const bodyText = todo.body?.trim() ? todo.body.trim() : "No details yet.";
+	const bodyLines = bodyText.split("\n");
+
+	const lines = [
+		summary,
+		theme.fg("muted", `Status: ${getTodoStatus(todo)}`),
+		theme.fg("muted", `Tags: ${tags}`),
+		theme.fg("muted", `Created: ${createdAt}`),
+		"",
+		theme.fg("muted", "Body:"),
+		...bodyLines.map((line) => theme.fg("text", `  ${line}`)),
+	];
+
+	return lines.join("\n");
+}
+
+export function appendExpandHint(theme: Theme, text: string): string {
+	return `${text}\n${theme.fg("dim", `(${keyHint("app.tools.expand", "to expand")})`)}`;
+}
